added Binary file checking

+++ Init.php
@@ -16,16 +16,17 @@ class Init
         if (strpos($os, 'win') !== false) {
             $baseDir = dirname(__FILE__). "\bin";
             
-            exec("powershell rm $baseDir\\textimg.exe");
+            if (file_exists("$baseDir\\textimg.exe")) exec("powershell rm $baseDir\\textimg.exe");
             // Windows
             echo "Detected Windows. Extracting textimg-windows-amd64.zip...\nThis may take a while...";
             exec("powershell Expand-Archive -Path $baseDir/textimg-windows-amd64.zip -DestinationPath $baseDir");
+            echo "\nExtract has completed";
         } else {
-            exec("rm -rf $baseDir/textimg");
-
+            if (file_exists("$baseDir\\textimg"))exec("rm -rf $baseDir/textimg");
             // Linux or other OS
             echo "Detected Linux. Extracting textimg-linux-amd64.tar.gz...\n";
             exec("tar -xzf $baseDir/textimg-linux-amd64.tar.gz -C $baseDir");
+            echo "\nExtract has completed";
         }
     }
 }
